Major progress in how requests are handled
Sets up an environment that delegates the work from a controller to a handler, which is free to throw exceptions and has complete freedom to return whatever it wants. Errors are handled automatically and properly logged. The handler registry works by filename and folder name as strings, the handlers receive the response object as their first argument then ...any[].